order_service/api/requirements.txt
Flask==3.0.0
Flask-CORS
mysql-connector-python==8.3.0
pika==1.3.2
psycopg[binary]
pandas
requests
